NexoraLearning/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
werkzeug>=3.0.0
